LayoutTests/http/tests/loading/resourceLoadStatistics/request-storage-access-same-origin-sandboxed-iframe-without-allow-token-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "theIframe" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "theIframe" - didCommitLoadForFrame
frame "theIframe" - didFinishDocumentLoadForFrame
CONSOLE MESSAGE: line 70: PASS document.hasStorageAccess was denied.
frame "theIframe" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "theIframe" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
Tests that same-origin iframe storage access is denied if the iframe is sandboxed but doesn't have the allow token.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
